refactor: extract rag_metadata_processing.py from rag_core (C-2, #557)
rag_core.py에서 metadata 처리 심볼들을 rag_metadata_processing.py로 추출.
rag_text_processing + rag_conversation_state만 의존하는 leaf에 가까운 모듈.

이동된 심볼 (23개):
Constants: WEAK_SECTION_HEADINGS, STRICT_METADATA_CONFIDENCE, REDUCED_METADATA_CONFIDENCE
Sections: normalize_regions, normalize_page_span, normalize_document_sections,
  document_has_section_structure, fixed_parent_section, split_section_text
Metadata targets: make_metadata_target, metadata_explicit_aliases, metadata_aliases
Matching: coerce_metadata_targets, match_metadata_targets, match_metadata_target,
  best_metadata_phrase_similarity
Match results: make_metadata_match, dedupe_metadata_matches, metadata_matches_for_stage,
  metadata_filters_from_matches, best_metadata_doc_scores, metadata_ambiguity_details
Also: metadata_tokens (was rag_core, now rag_metadata_processing)

late-import → module-level 전환:
- rag_verifier: metadata_tokens
- rag_retrieval: normalize_regions, normalize_page_span (2곳)
- rag_answer: normalize_page_span, normalize_regions
- rag_query: 8개 심볼 + metadata_tokens

rag_core backward-compat re-export 유지.

Closes #557
